requirements.txt
fastapi
uvicorn
httpx
python-dotenv
motor
PyJWT